configs | cleanup - remove/rename redundant code
