+++ configs/train_epic_RGB.yaml
