+++ original/script_test_sourceonly.sh
